Kokkos: Fixed a bug whereby I was casting a Scalar to a LocalOrdinal directly, rather than taking the reeal part of the Scalar value first.  The Scalar might be complex.  (It comes from an LAPACK LWORK query, so the real part of the Scalar is safe to cast to the ordinal type used by the LAPACK implementation.) Belos: Fixed a compiler warning about comparing a signed and unsigned integer. Tpetra: Fixed a syntax error by removing an unnecessary "typename" keyword.  This only affects the Tpetra unit tests, not the actual Tpetra object declarations or definitions.
Build/Test Cases Summary
Enabled Packages: Belos, Kokkos, Tpetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=542,notpassed=0 (53.22 min)
1) SERIAL_RELEASE => passed: passed=514,notpassed=0 (30.90 min)

+++ packages/belos/tpetra/test/OrthoManager/belos_orthomanager_tpetra.cpp
@@ -494,7 +494,10 @@ main (int argc, char *argv[])
     // sparse matrix is successfully loaded, its number of rows will
     // override the number of rows specified on the command line (if
     // specified), and will also override the default number of rows.
-    const int maxNormalizeNumCols = std::max (sizeS, std::max (sizeX1, sizeX2));
+    const size_t maxNormalizeNumCols = std::max (sizeS, std::max (sizeX1, sizeX2));
+    // getNodeNumElements() returns a size_t, which is unsigned, and
+    // you shouldn't compare signed and unsigned values.  This is why
+    // we make maxNormalizeNumCols a size_t as well.
     if (map->getNodeNumElements() < maxNormalizeNumCols)
       {
 	std::ostringstream os;
